Added static client store
